% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean_exp}
\alias{euclidean_exp}
\title{Compute the Euclidean Exponential}
\usage{
euclidean_exp(sigma, v, validate = FALSE)
}
\arguments{
\item{sigma}{A reference point.}

\item{v}{A tangent vector to be mapped back to the manifold at \code{sigma}.}

\item{validate}{A logical value indicating whether to validate input arguments. Default is FALSE.}
}
\value{
The point on the manifold corresponding to the tangent vector at \code{sigma}.
}
\description{
Compute the Euclidean Exponential
}
