% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean_log}
\alias{euclidean_log}
\title{Compute the Euclidean Logarithm}
\usage{
euclidean_log(sigma, lambda, validate = FALSE)
}
\arguments{
\item{sigma}{A reference point.}

\item{lambda}{A point on the manifold.}

\item{validate}{A logical value indicating whether to validate input arguments. Default is FALSE.}
}
\value{
The tangent space image of \code{lambda} at \code{sigma}.
}
\description{
Compute the Euclidean Logarithm
}
