% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_config.R
\name{get_n_workers}
\alias{get_n_workers}
\title{Get Current Number of Parallel Workers}
\usage{
get_n_workers()
}
\value{
Integer specifying the number of workers, or 1 if sequential processing is enabled.
}
\description{
Returns the number of parallel workers configured in the current future plan.
}
\examples{
\dontrun{
set_parallel_plan("multisession", workers = 4)
get_n_workers()  # Returns 4

set_parallel_plan("sequential")
get_n_workers()  # Returns 1
}

}
\seealso{
\code{\link[=set_parallel_plan]{set_parallel_plan()}}
}
