% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemtan-package.R
\docType{package}
\name{riemtan-package}
\alias{riemtan}
\alias{riemtan-package}
\title{riemtan: Riemannian Metrics for Symmetric Positive Definite Matrices}
\description{
Implements various Riemannian metrics for symmetric positive definite matrices, including AIRM (Affine Invariant Riemannian Metric, \doi{10.1007/s11263-005-3222-z}), Log-Euclidean (\doi{10.1002/mrm.20965}), Euclidean, Log-Cholesky (\doi{10.1137/18M1221084}), and Bures-Wasserstein metrics (\doi{10.1016/j.exmath.2018.01.002}). Provides functions for computing logarithmic and exponential maps, vectorization, and statistical operations on the manifold of positive definite matrices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nicoesve.github.io/riemtan/}
  \item Report bugs at \url{https://github.com/nicoesve/riemtan/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Escobar \email{nescoba@iu.edu} (\href{https://orcid.org/0009-0006-0800-5692}{ORCID})

Other contributors:
\itemize{
  \item Jaroslaw Harezlak [thesis advisor]
}

}
\keyword{internal}
